add module to store structs,enums,...
